requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
docopt==0.6.2
fabric @ git+https://github.com/Fabric-Development/fabric.git@8633df172a3ceee9222e7e583e93717f733d5618
idna==3.11
loguru==0.7.3
psutil==7.2.0
pulsectl==24.12.0
py-mon==2.2.0
pycairo==1.29.0
PyGObject==3.50.0
requests==2.32.5
urllib3==2.6.2
watchdog==6.0.0
yarg==0.1.9
